+++ packages/common/src/index.ts
@@ -225,6 +225,10 @@ export * from './utils/slugs';
 export * from './utils/time';
 export * from './utils/timeFrames';
 export * from './utils/warehouse';
+export * from './visualizations/BarChartDataTransformer';
+export * from './visualizations/ResultsTransformerBase';
+export * from './visualizations/SqlRunnerResultsTableTransformer';
+export * from './visualizations/SqlRunnerResultsTransformer';
 
 export const validateEmail = (email: string): boolean => {
     if (/\s/.test(email)) {