+++ packages/e2e/cypress/e2e/api/sqlRunner.cy.ts
@@ -264,7 +264,7 @@ describe.skip(`Load testing`, () => {
     });
 });
 
-describe.only(`Saved SQL chart`, () => {
+describe(`Saved SQL chart`, () => {
     beforeEach(() => {
         cy.login();
     });