+++ packages/frontend/src/features/sqlRunner/components/ContentPanel.tsx
@@ -94,8 +94,8 @@ export const ContentPanel: FC<Props> = ({
         isLoading,
     } = useSqlQueryRun({
         onSuccess: (data) => {
-            if (data?.results) {
-                dispatch(setInitialResultsAndSeries(data.results));
+            if (data) {
+                dispatch(setInitialResultsAndSeries(data));
                 if (resultsHeight === MIN_RESULTS_HEIGHT) {
                     setResultsHeight(inputSectionHeight / 2);
                 }
@@ -350,12 +350,13 @@ export const ContentPanel: FC<Props> = ({
                                             />
                                         )}
                                         {selectedChartType ===
-                                            ChartKind.VERTICAL_BAR && (
-                                            <BarChart
-                                                data={queryResults.results}
-                                                config={barChartConfig}
-                                            />
-                                        )}
+                                            ChartKind.VERTICAL_BAR &&
+                                            barChartConfig && (
+                                                <BarChart
+                                                    data={queryResults}
+                                                    config={barChartConfig}
+                                                />
+                                            )}
                                     </>
                                 )}
                                 {activeVisTab === VisTabs.RESULTS && (