+++ packages/frontend/src/features/sqlRunner/components/RightSidebar.tsx
@@ -2,6 +2,7 @@ import { ChartKind } from '@lightdash/common';
 import {
     ActionIcon,
     Group,
+    MantineProvider,
     SegmentedControl,
     Stack,
     Title,
@@ -10,9 +11,11 @@ import {
 import { IconLayoutSidebarRightCollapse } from '@tabler/icons-react';
 import { type Dispatch, type FC, type SetStateAction } from 'react';
 import MantineIcon from '../../../components/common/MantineIcon';
+
+import { themeOverride } from '../../../components/VisualizationConfigs/mantineTheme';
 import { useAppDispatch, useAppSelector } from '../store/hooks';
 import { setSelectedChartType } from '../store/sqlRunnerSlice';
-import { default as BarChartConfiguration } from './BarChartConfiguration';
+import { BarChartConfig } from './BarChartConfiguration';
 import TableVisConfiguration from './TableVisConfiguration';
 
 type Props = {
@@ -27,37 +30,41 @@ export const RightSidebar: FC<Props> = ({ setSidebarOpen }) => {
     );
 
     return (
-        <Stack h="100vh" spacing="xs">
-            <Group position="apart">
-                <Title order={5} fz="sm" c="gray.6">
-                    Configure Chart
-                </Title>
-                <Tooltip variant="xs" label="Close sidebar" position="left">
-                    <ActionIcon size="xs">
-                        <MantineIcon
-                            icon={IconLayoutSidebarRightCollapse}
-                            onClick={() => setSidebarOpen(false)}
-                        />
-                    </ActionIcon>
-                </Tooltip>
-            </Group>
+        <MantineProvider inherit theme={themeOverride}>
+            <Stack h="100vh" spacing="xs">
+                <Group position="apart">
+                    <Title order={5} fz="sm" c="gray.6">
+                        Configuration
+                    </Title>
+                    <Tooltip variant="xs" label="Close sidebar" position="left">
+                        <ActionIcon size="xs">
+                            <MantineIcon
+                                icon={IconLayoutSidebarRightCollapse}
+                                onClick={() => setSidebarOpen(false)}
+                            />
+                        </ActionIcon>
+                    </Tooltip>
+                </Group>
 
-            <SegmentedControl
-                size="xs"
-                value={selectedChartType}
-                onChange={(value: ChartKind) =>
-                    dispatch(setSelectedChartType(value))
-                }
-                data={[
-                    { value: ChartKind.TABLE, label: 'Table' },
-                    { value: ChartKind.VERTICAL_BAR, label: 'Bar chart' },
-                ]}
-            />
+                <SegmentedControl
+                    size="xs"
+                    value={selectedChartType}
+                    onChange={(value: ChartKind) =>
+                        dispatch(setSelectedChartType(value))
+                    }
+                    data={[
+                        { value: ChartKind.TABLE, label: 'Table' },
+                        { value: ChartKind.VERTICAL_BAR, label: 'Bar chart' },
+                    ]}
+                />
 
-            {selectedChartType === ChartKind.TABLE && <TableVisConfiguration />}
-            {selectedChartType === ChartKind.VERTICAL_BAR && (
-                <BarChartConfiguration />
-            )}
-        </Stack>
+                {selectedChartType === ChartKind.TABLE && (
+                    <TableVisConfiguration />
+                )}
+                {selectedChartType === ChartKind.VERTICAL_BAR && (
+                    <BarChartConfig />
+                )}
+            </Stack>
+        </MantineProvider>
     );
 };