+++ packages/frontend/src/features/sqlRunner/components/TableFields.tsx
@@ -34,7 +34,7 @@ import {
 } from '../hooks/useTableFields';
 import { useAppSelector } from '../store/hooks';
 
-const TableFieldIcon: FC<{
+export const TableFieldIcon: FC<{
     fieldType: DimensionType;
 }> = memo(({ fieldType }) => {
     const Icon = useMemo(() => {