+++ packages/frontend/src/features/sqlRunner/store/barChartSlice.ts
@@ -1,68 +1,158 @@
-import { ChartKind, type BarChartConfig } from '@lightdash/common';
+import {
+    ChartKind,
+    type AggregationOptions,
+    type BarChartConfig,
+    type BarChartDisplay,
+    type XLayoutOptions,
+    type YLayoutOptions,
+} from '@lightdash/common';
 import type { PayloadAction } from '@reduxjs/toolkit';
 import { createSlice } from '@reduxjs/toolkit';
+import { SqlRunnerResultsTransformerFE } from '../transformers/useBarChart';
 import { setInitialResultsAndSeries, setSaveChartData } from './sqlRunnerSlice';
 
-const initialState: { config: BarChartConfig | undefined } = {
+type InitialState = {
+    config: BarChartConfig | undefined;
+    options: {
+        xLayoutOptions: XLayoutOptions[];
+        yLayoutOptions: YLayoutOptions[];
+    };
+};
+
+const initialState: InitialState = {
     config: undefined,
+    options: {
+        xLayoutOptions: [],
+        yLayoutOptions: [],
+    },
 };
 
 export const barChartConfigSlice = createSlice({
     name: 'barChartConfig',
     initialState,
     reducers: {
-        setXAxisLabel: ({ config }, action: PayloadAction<string>) => {
-            if (config?.axes) {
-                config.axes.x.label = action.payload;
+        // TODO: add y field feature
+        // addYField: (
+        //     state,
+        //     action: PayloadAction<SqlTransformBarChartConfig['y'][number]>,
+        // ) => {
+        //     if (!state.config) return;
+        //     if (!state.config.fieldConfig) return;
+
+        //     const yAxisFields = state.config.fieldConfig.y;
+        //     if (yAxisFields) {
+        //         state.config.fieldConfig.y = [...yAxisFields, action.payload];
+        //     }
+        // },
+        setXAxisReference: ({ config }, action: PayloadAction<string>) => {
+            if (config?.fieldConfig?.x) {
+                config.fieldConfig.x.reference = action.payload;
             }
         },
-        setYAxisLabel: ({ config }, action: PayloadAction<string>) => {
-            if (config?.axes) {
-                config.axes.y[0].label = action.payload;
+        setYAxisReference: (
+            { config },
+            action: PayloadAction<{
+                previousReference: string;
+                reference: string;
+                aggregation: AggregationOptions;
+            }>,
+        ) => {
+            if (config?.fieldConfig?.y) {
+                const yAxis = config.fieldConfig.y.find(
+                    (axis) =>
+                        axis.reference === action.payload.previousReference,
+                );
+                if (yAxis) {
+                    yAxis.reference = action.payload.reference;
+                    yAxis.aggregation = action.payload.aggregation;
+                }
+            }
+        },
+        setYAxisAggregation: (
+            { config },
+            action: PayloadAction<{
+                reference: string;
+                aggregation: AggregationOptions;
+            }>,
+        ) => {
+            if (!config) return;
+            if (!config?.fieldConfig?.y) return;
+
+            const yAxis = config.fieldConfig.y.find(
+                (axis) => axis.reference === action.payload.reference,
+            );
+            if (yAxis) {
+                yAxis.aggregation = action.payload.aggregation;
+            }
+        },
+        setXAxisLabel: (
+            { config },
+            action: PayloadAction<BarChartDisplay['xAxis']>,
+        ) => {
+            if (!config) return;
+            if (!config.display) {
+                config.display = {
+                    xAxis: action.payload,
+                };
+            } else {
+                config.display.xAxis = action.payload;
+            }
+        },
+        setYAxisLabel: (
+            { config },
+            action: PayloadAction<{ index: number; label: string }>,
+        ) => {
+            if (!config) return;
+            if (!config.display) {
+                config.display = {
+                    yAxis: [
+                        {
+                            label: action.payload.label,
+                        },
+                    ],
+                };
+            }
+            if (config?.display?.yAxis) {
+                config.display.yAxis[action.payload.index].label =
+                    action.payload.label;
             }
         },
         setSeriesLabel: (
             { config },
-            action: PayloadAction<{ index: number; label: string }>,
+            action: PayloadAction<{
+                reference: string;
+                label: string;
+            }>,
         ) => {
-            if (config?.series) {
-                config.series[action.payload.index].name = action.payload.label;
+            if (!config || !config.display) return;
+
+            if (config.display.series) {
+                config.display.series[action.payload.reference].label =
+                    action.payload.label;
             }
         },
     },
     extraReducers: (builder) => {
         builder.addCase(setInitialResultsAndSeries, (state, action) => {
-            if (state.config === undefined) {
-                const fieldIds = Object.keys(action.payload[0]);
-                state.config = {
-                    metadata: {
-                        version: 1,
-                    },
-                    type: ChartKind.VERTICAL_BAR,
-                    style: {
-                        legend: {
-                            position: 'top',
-                            align: 'center',
-                        },
-                    },
-                    axes: {
-                        x: {
-                            reference: fieldIds[0],
-                            label: fieldIds[0],
-                        },
-                        y: [
-                            {
-                                reference: fieldIds[1],
-                                label: fieldIds[1],
-                            },
-                        ],
-                    },
-                    series: fieldIds.slice(1).map((reference, index) => ({
-                        reference,
-                        name: reference,
-                        yIndex: index,
-                    })),
-                };
+            if (
+                state.config === undefined &&
+                action.payload.results &&
+                action.payload.columns
+            ) {
+                const sqlRunnerResultsTransformer =
+                    new SqlRunnerResultsTransformerFE({
+                        rows: action.payload.results,
+                        columns: action.payload.columns,
+                    });
+                if (action.payload.columns) {
+                    state.options = {
+                        xLayoutOptions:
+                            sqlRunnerResultsTransformer.barChartXLayoutOptions(),
+                        yLayoutOptions:
+                            sqlRunnerResultsTransformer.barChartYLayoutOptions(),
+                    };
+                    state.config = sqlRunnerResultsTransformer.defaultConfig();
+                }
             }
         });
         builder.addCase(setSaveChartData, (state, action) => {
@@ -73,5 +163,11 @@ export const barChartConfigSlice = createSlice({
     },
 });
 
-export const { setXAxisLabel, setYAxisLabel, setSeriesLabel } =
-    barChartConfigSlice.actions;
+export const {
+    setYAxisLabel,
+    setXAxisLabel,
+    setXAxisReference,
+    setYAxisReference,
+    setYAxisAggregation,
+    setSeriesLabel,
+} = barChartConfigSlice.actions;