+++ packages/frontend/src/features/sqlRunner/store/sqlRunnerSlice.ts
@@ -1,13 +1,13 @@
 import {
     ChartKind,
-    type ResultRow,
     type SqlChart,
+    type SqlColumn,
     type SqlTableConfig,
     type TableChartSqlConfig,
 } from '@lightdash/common';
-
 import type { PayloadAction } from '@reduxjs/toolkit';
 import { createSlice } from '@reduxjs/toolkit';
+import { type ResultsAndColumns } from '../hooks/useSqlQueryRun';
 
 export enum VisTabs {
     CHART = 'chart',
@@ -42,6 +42,8 @@ export interface SqlRunnerState {
     };
 
     quoteChar: string;
+
+    sqlColumns: SqlColumn[] | undefined;
 }
 
 const initialState: SqlRunnerState = {
@@ -61,6 +63,7 @@ const initialState: SqlRunnerState = {
         },
     },
     quoteChar: '"',
+    sqlColumns: undefined,
 };
 
 export const sqlRunnerSlice = createSlice({
@@ -75,11 +78,13 @@ export const sqlRunnerSlice = createSlice({
         },
         setInitialResultsAndSeries: (
             state,
-            action: PayloadAction<ResultRow[]>,
+            action: PayloadAction<ResultsAndColumns>,
         ) => {
-            // TODO: this should come from the transformer
+            if (!action.payload.results || !action.payload.columns) {
+                return;
+            }
             // Set the initial results table config
-            const columns = Object.keys(action.payload[0]).reduce<
+            const columns = Object.keys(action.payload.results[0]).reduce<
                 TableChartSqlConfig['columns']
             >(
                 (acc, key) => ({